+++ genie-common/src/main/java/com/netflix/genie/common/dto/RuntimeResources.java
@@ -41,7 +41,7 @@ public class RuntimeResources implements Serializable {
     private final Integer gpu;
 
     @Min(value = 1, message = "Must have at least 1 MB of memory")
-    private final Long memoryMb;
+    private final Integer memoryMb;
 
     @Min(value = 1, message = "Must have at least 1 MB of disk space")
     private final Long diskMb;
@@ -80,7 +80,7 @@ public class RuntimeResources implements Serializable {
      *
      * @return The amount or {@link Optional#empty()}
      */
-    public Optional<Long> getMemoryMb() {
+    public Optional<Integer> getMemoryMb() {
         return Optional.ofNullable(this.memoryMb);
     }
 
@@ -153,7 +153,7 @@ public class RuntimeResources implements Serializable {
 
         private Integer bCpu;
         private Integer bGpu;
-        private Long bMemoryMb;
+        private Integer bMemoryMb;
         private Long bDiskMb;
         private Long bNetworkMbps;
 
@@ -185,7 +185,7 @@ public class RuntimeResources implements Serializable {
          * @param memoryMb The number must be at least 1 or {@literal null}
          * @return The {@link Builder}
          */
-        public Builder withMemoryMb(@Nullable final Long memoryMb) {
+        public Builder withMemoryMb(@Nullable final Integer memoryMb) {
             this.bMemoryMb = memoryMb;
             return this;
         }
